Move ErrorStack and StackFrame to ScynamoError file
